+++ ncc_page.py
@@ -33,7 +33,7 @@ import os
 import configparser
 from pathlib import Path
 from newapi.super import bot_api
-from newapi.super import super_page
+from newapi.super.S_Page import super_page
 from newapi.super.S_Category import catdepth_new
 
 print_test = {1: True if "test" in sys.argv else False}
