+++ page.py
@@ -27,7 +27,7 @@ from newapi.page import MainPage, NEW_API
 import os
 import sys
 from newapi.super import bot_api
-from newapi.super import super_page
+from newapi.super.S_Page import super_page
 from newapi.super.S_Category import catdepth_new
 
 from newapi import useraccount
