+++ toolforge_page.py
@@ -14,7 +14,7 @@ MainPage = toolforge_page.super_page.MainPage
 # ---
 """
 # ---
-from newapi.super import super_page
+from newapi.super.S_Page import super_page
 from newapi.super.S_Category import catdepth_new
 
 __all__ = [
